Add option for more efficient, alternative hopper protection
This disables protection checks for any item moves except for hopper minecarts. Protection against hoppers is now done by checking for protections of the containers that the hopper is pointing into or that are above the hopper when the player places a hopper. If the player doesn't have access to it or the hopper flag isn't allowing it then the placement is blocked.