Fix bindMessageCache not expiring values
Fixes #130
This was a serious flaw in the simple message parser. Not sure
how it wasn't caught by now, but thanks to @aurorasmiles we know
now that it's been a problem. It now uses an LRU cache with a
maximum capacity of 1000. This may be a bit small, so I've also
added cache hits/misses to '/lwc debug' as a way to check whether
this will need to be increased.