Revert "Don't match door if neither door block is protected"
This reverts commit 4acf69e627e0118ad68f3e1727cabd9d6e8cc2cd.
It was causing strange issues with doors not being properly
protected for both sides of the block in some instances,
although it's unclear what the cause was. Since this was
only added as a workaround for a rare issue with limits,
it's probably better to come up with another solution for
that and revert this to make sure there aren't any further
problems.