Remove footgun with primary key automatically setting autoincrement.
They are two separate concepts, and it shouldn't be taken for granted
that you need to unset the autoincrement after setting isPrimary to
true for non integer datatypes. Especially because this works just fine
in sqlite, but not in mysql, meaning that it's likely to work just fine
in testing, but not in prod.